feat: update token generator to include roles, permissions and email claims
