+++ src/InvenEase.Infrastructure/Security/TokenGenerator/JwtTokenGenerator.cs
@@ -1,0 +1,49 @@
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using System.Text;
+
+using InvenEase.Application.Common.Interfaces.Authentication;
+using InvenEase.Application.Common.Interfaces.Services;
+using InvenEase.Domain.UserAggregate;
+
+using Microsoft.Extensions.Options;
+using Microsoft.IdentityModel.Tokens;
+
+namespace InvenEase.Infrastructure.Security.TokenGenerator;
+
+public class JwtTokenGenerator(IDateTimeProvider dateTimeProvider, IOptions<JwtSettings> jwtOptions) : IJwtTokenGenerator
+{
+    private readonly JwtSettings _jwtSettings = jwtOptions.Value;
+
+    public string GenerateToken(User user)
+    {
+        var signingCredentials = new SigningCredentials(
+            new SymmetricSecurityKey(
+                Encoding.UTF8.GetBytes(_jwtSettings.Secret)), SecurityAlgorithms.HmacSha256);
+
+        var claims = new List<Claim>
+        {
+            new(JwtRegisteredClaimNames.Sub, user.Id.Value.ToString()),
+            new(JwtRegisteredClaimNames.GivenName, user.FirstName),
+            new(JwtRegisteredClaimNames.FamilyName, user.LastName),
+            new(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()),
+            new(ClaimTypes.Email, user.Email),
+        };
+
+        user.Roles
+            .ToList()
+            .ForEach(role => claims.Add(new(ClaimTypes.Role, role.Value)));
+        user.Permissions
+            .ToList()
+            .ForEach(permission => claims.Add(new("permissions", permission.Value)));
+
+        var securityToken = new JwtSecurityToken(
+            issuer: _jwtSettings.Issuer,
+            audience: _jwtSettings.Audience,
+            expires: dateTimeProvider.UtcNow.AddMinutes(_jwtSettings.ExpiryMinutes),
+            claims: claims,
+            signingCredentials: signingCredentials);
+
+        return new JwtSecurityTokenHandler().WriteToken(securityToken);
+    }
+}
